Update lots of stuff
Tests are now back in working fashion and work concurrently now
(which the issues were caused by good ol' dshow properties).

Added a new completely dependent test which does nothing but
stream to my channel with a color source and a desktop capture
source. The point of this was to prove the API works.

There are no extensive tests for outputs/services/encoders yet.